Fix crash in PerProfileWorkerTaskTracker::CreateWorkerTask()
Because service worker notifications can arrive asynchronously, their
render process host could already have been deleted.

With this change, those service workers are ignored to avoid calling
into a dead RPH.

One of the goals with this CL is to be minimal enough so that it can be
easily reverted when service worker notifications are fixed to arrive
synchronously.

Fixed: 1191930
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2823169
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/master@{#872119}

+++ chrome/browser/task_manager/providers/per_profile_worker_task_tracker.cc
@@ -62,8 +62,11 @@ void PerProfileWorkerTaskTracker::OnWorkerCreated(
     const blink::DedicatedWorkerToken& worker_token,
     int worker_process_id,
     content::GlobalFrameRoutingId ancestor_render_frame_host_id) {
+  auto* worker_process_host =
+      content::RenderProcessHost::FromID(worker_process_id);
+  DCHECK(worker_process_host);
   CreateWorkerTask(worker_token, Task::Type::DEDICATED_WORKER,
-                   worker_process_id, &dedicated_worker_tasks_);
+                   worker_process_host, &dedicated_worker_tasks_);
 }
 
 void PerProfileWorkerTaskTracker::OnBeforeWorkerDestroyed(
@@ -82,8 +85,11 @@ void PerProfileWorkerTaskTracker::OnWorkerCreated(
     const blink::SharedWorkerToken& shared_worker_token,
     int worker_process_id,
     const base::UnguessableToken& dev_tools_token) {
+  auto* worker_process_host =
+      content::RenderProcessHost::FromID(worker_process_id);
+  DCHECK(worker_process_host);
   CreateWorkerTask(shared_worker_token, Task::Type::SHARED_WORKER,
-                   worker_process_id, &shared_worker_tasks_);
+                   worker_process_host, &shared_worker_tasks_);
 }
 
 void PerProfileWorkerTaskTracker::OnBeforeWorkerDestroyed(
@@ -100,13 +106,33 @@ void PerProfileWorkerTaskTracker::OnFinalResponseURLDetermined(
 void PerProfileWorkerTaskTracker::OnVersionStartedRunning(
     int64_t version_id,
     const content::ServiceWorkerRunningInfo& running_info) {
-  CreateWorkerTask(version_id, Task::Type::SERVICE_WORKER,
-                   running_info.render_process_id, &service_worker_tasks_);
+  auto* worker_process_host =
+      content::RenderProcessHost::FromID(running_info.render_process_id);
+
+  // It's possible that the renderer is already gone since the notification for
+  // a service worker comes asynchronously. Ignore this worker.
+  if (!worker_process_host) {
+    // A matching OnVersionStoppedRunning() call is still expected for this
+    // service worker.
+    const bool inserted = ignored_service_worker_.insert(version_id).second;
+    DCHECK(inserted);
+    return;
+  }
+
+  CreateWorkerTask(version_id, Task::Type::SERVICE_WORKER, worker_process_host,
+                   &service_worker_tasks_);
   SetWorkerTaskScriptUrl(version_id, running_info.script_url,
                          &service_worker_tasks_);
 }
 
 void PerProfileWorkerTaskTracker::OnVersionStoppedRunning(int64_t version_id) {
+  size_t removed = ignored_service_worker_.erase(version_id);
+  if (removed) {
+    // A task for this service worker was never created. Ignore the
+    // notification.
+    return;
+  }
+
   DeleteWorkerTask(version_id, &service_worker_tasks_);
 }
 
@@ -114,14 +140,13 @@ template <typename WorkerId>
 void PerProfileWorkerTaskTracker::CreateWorkerTask(
     const WorkerId& worker_id,
     Task::Type task_type,
-    int worker_process_id,
+    content::RenderProcessHost* worker_process_host,
     base::flat_map<WorkerId, std::unique_ptr<WorkerTask>>* out_worker_tasks) {
-  auto* worker_process_host =
-      content::RenderProcessHost::FromID(worker_process_id);
+  DCHECK(worker_process_host);
   auto insertion_result = out_worker_tasks->emplace(
       worker_id,
       std::make_unique<WorkerTask>(worker_process_host->GetProcess().Handle(),
-                                   task_type, worker_process_id));
+                                   task_type, worker_process_host->GetID()));
   DCHECK(insertion_result.second);
   worker_task_provider_->OnWorkerTaskAdded(
       insertion_result.first->second.get());
